{{ cookiecutter.project_slug }}/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
pyyaml==6.0.2
requests==2.32.3
    # via requests-oauthlib
requests-oauthlib==2.0.0
ruff==0.5.7
urllib3==2.2.2
    # via requests
